feat(LottoController.java): 시스템 흐름에 맞게 Model과 View를 배치
OuputView에서 일치하지 않았던 타입을 동일하게 변경하였다.
그리고 InputView에서 사용자가 숫자를 입력하는 부분을 구현하였음.

+++ src/main/java/lotto/View/OutputView.java
@@ -18,7 +18,7 @@ public class OutputView {
         });
     }
 
-    public void printWinningCounts(Map<LottoGame.LottoPrice, Integer> winningCounts) {
+    public void printWinningCounts(Map<LottoGame.LottoPrice, Long> winningCounts) {
         System.out.println("당첨 통계\n---");
         List<LottoGame.LottoPrice> winnings = List.of(
                 LottoGame.LottoPrice.FIFTH,
